+++ src/core/proto/index.ts
@@ -1,5 +1,0 @@
-export * from './extra'
-
-export * from './token-key'
-export * from './token-key-create'
-export * from './token-key-list-query'