+++ src/helpers/setup/swagger.ts
@@ -6,8 +6,6 @@ import {
   type SwaggerCustomOptions,
 } from '@nestjs/swagger'
 
-import { PaginationQuery } from '~/core/proto'
-
 export const setupSwagger = (path: string, app: INestApplication) => {
   const configService = app.get<ConfigService>(ConfigService)
 
@@ -23,12 +21,9 @@ export const setupSwagger = (path: string, app: INestApplication) => {
     )
     .setVersion(configService.get('version'))
     .addTag('Auth')
-    .addTag('Tokens')
     .build()
 
-  const document = SwaggerModule.createDocument(app, config, {
-    extraModels: [PaginationQuery],
-  })
+  const document = SwaggerModule.createDocument(app, config)
 
   const options: SwaggerCustomOptions = {
     swaggerOptions: {